flask-backend/requirements.txt
Authlib==1.0.0
base58==2.1.1
cryptography==35.0.0
ecdsa==0.17.0
filelock==3.4.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Dance==5.1.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.8.1
python-decouple==3.5
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
urllib3==1.26.7
virtualenv==20.10.0
geocoder
flask_session_captcha
flask_sessionstore
Pillow
PyJWT==2.3.0
ecdsa

